Problem: parser does not parse JSON content
